# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Conference Management System API",
    "description": "API for managing conferences, attendees, sessions, and more",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@conference.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/permissions": {
      "get": {
        "summary": "List permissions",
        "description": "Get paginated list of permissions",
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create permission",
        "description": "Create a new permission",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code",
                "examples": [
                  "VALIDATION_ERROR",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "description": "Field name with error"
                    },
                    "message": {
                      "type": "string",
                      "description": "Field-specific error message"
                    }
                  }
                },
                "description": "Detailed error information"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "integer",
            "description": "Unique identifier for the permission"
          },
          "CODE": {
            "type": "string",
            "description": "Permission code",
            "maxLength": 50
          },
          "NAME": {
            "type": "string",
            "description": "Permission name",
            "maxLength": 100
          },
          "CATEGORY": {
            "type": "string",
            "description": "Permission category",
            "maxLength": 50
          },
          "DESCRIPTION": {
            "type": "string",
            "description": "Permission description",
            "maxLength": 255
          }
        },
        "required": [
          "CODE",
          "NAME"
        ]
      },
      "CreatePermissionRequest": {
        "type": "object",
        "properties": {
          "CODE": {
            "type": "string",
            "description": "Permission code",
            "maxLength": 50
          },
          "NAME": {
            "type": "string",
            "description": "Permission name",
            "maxLength": 100
          },
          "CATEGORY": {
            "type": "string",
            "description": "Permission category",
            "maxLength": 50
          },
          "DESCRIPTION": {
            "type": "string",
            "description": "Permission description",
            "maxLength": 255
          }
        },
        "required": [
          "CODE",
          "NAME"
        ]
      },
      "PermissionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Permission"
          }
        }
      },
      "PermissionListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "description": "Current page number"
              },
              "limit": {
                "type": "integer",
                "description": "Items per page"
              },
              "total": {
                "type": "integer",
                "description": "Total number of items"
              },
              "totalPages": {
                "type": "integer",
                "description": "Total number of pages"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}